backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
databases==0.9.0
fastapi==0.117.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
lxml==6.0.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyPDF2==3.0.1
pypdfium2==4.30.0
python-docx==1.2.0
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
